fix: address PR comment, move factories
