+++ packages/network/test/integration/webrtc-endpoint-back-pressure-handling.test.ts
@@ -9,7 +9,7 @@ import { TrackerNode } from '../../src/protocol/TrackerNode'
 import { wait } from 'streamr-test-utils'
 import { NegotiatedProtocolVersions } from "../../src/connection/NegotiatedProtocolVersions"
 import { WebRtcEndpoint } from '../../src/connection/WebRtcEndpoint'
-import { NodeWebRtcConnectionFactory } from '../../src/connection/webRtcConnectionFactories'
+import { NodeWebRtcConnectionFactory } from "../../src/connection/NodeWebRtcConnection"
 
 describe('WebRtcEndpoint: back pressure handling', () => {
     let tracker: Tracker
