+++ packages/network/test/integration/webrtc-multi-signaller.test.ts
@@ -8,7 +8,7 @@ import { Event as EndpointEvent } from '../../src/connection/IWebRtcEndpoint'
 import { RtcSignaller } from '../../src/logic/RtcSignaller'
 import { NegotiatedProtocolVersions } from "../../src/connection/NegotiatedProtocolVersions"
 import { WebRtcEndpoint } from '../../src/connection/WebRtcEndpoint'
-import { NodeWebRtcConnectionFactory } from '../../src/connection/webRtcConnectionFactories'
+import { NodeWebRtcConnectionFactory } from "../../src/connection/NodeWebRtcConnection"
 
 describe('WebRTC multisignaller test', () => {
     let tracker1: Tracker
